+++ docs/fr_FR/changelog.md
@@ -40,5 +40,7 @@ Add 15 octobre 2020
 ---
 
 -   Ajout de la température des gaz
+-   Ajout de la valeur de puissance
+-   Ajout de la valeur de ventilation
 
 [Documentation](index.md)